requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy>=1.24.0
protobuf<5.0.0
pygame-ce>=2.4.0